+++ packages/ui5-test-writer/templates/v4/integration/ObjectPageJourney.js
@@ -62,13 +62,19 @@ sap.ui.define([
 
 <% if (bodySections?.length > 0) { -%>
         opaTest("Check body sections of the Object Page", function (Given, When, Then) {
+<% if (bodySections?.length > 1) { -%>
             Then.onThe<%- name%>.iCheckNumberOfSections(<%- bodySections.length %>);
+<% } -%>
 <% bodySections.forEach(function(section) { -%>
-            When.onThe<%- name%>.iGoToSection({ section: "<%- section.id %>" });
+<% if (!isStandalone) { -%>
+<% if (bodySections.length > 1) { -%>
+            When.onThe<%- name%>.iPressSectionIconTabFilterButton("<%- section.id %>");
+<% } -%>
             Then.onThe<%- name%>.iCheckSection({ section: "<%- section.id %>" });
+<% } -%>
 <% if (section?.subSections?.length > 0) { -%>
 <% section.subSections.forEach(function(subSection) { -%>
-            When.onThe<%- name%>.iGoToSection({ section: "<%- section.id %>", subSection: "<%- subSection.id %>" });
+            //When.onThe<%- name%>.iGoToSection({ section: "<%- section.id %>", subSection: "<%- subSection.id %>" });
             Then.onThe<%- name%>.iCheckSubSection({ section: "<%- subSection.id %>" });
 <% }) -%>
 <% } -%>