+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.1.135
+
+### Patch Changes
+
+- @sap-ux/fe-fpm-writer@0.43.21
+
+## 0.1.134
+
+### Patch Changes
+
+- Updated dependencies [9700a95]
+    - @sap-ux/fe-fpm-writer@0.43.20
+
 ## 0.1.133
 
 ### Patch Changes