+++ packages/cf-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 0.5.119
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.11.37
+
+## 0.5.118
+
+### Patch Changes
+
+- Updated dependencies [ee68603]
+    - @sap-ux/btp-utils@1.1.14
+    - @sap-ux/inquirer-common@0.11.36
+
+## 0.5.117
+
+### Patch Changes
+
+- Updated dependencies [cc4450c]
+    - @sap-ux/btp-utils@1.1.13
+    - @sap-ux/inquirer-common@0.11.35
+
 ## 0.5.116
 
 ### Patch Changes