+++ packages/nodejs-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/nodejs-utils
 
+## 0.2.21
+
+### Patch Changes
+
+- Updated dependencies [ee68603]
+    - @sap-ux/btp-utils@1.1.14
+
+## 0.2.20
+
+### Patch Changes
+
+- Updated dependencies [cc4450c]
+    - @sap-ux/btp-utils@1.1.13
+
 ## 0.2.19
 
 ### Patch Changes